Move BoarsSquares class to separate module
